pyproject.toml
[project]
name = "flash-cs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "chromadb-client>=1.0.5",
    "langchain-community>=0.3.21",
    "langchain>=0.3.23",
    "matplotlib>=3.9.4",
    "pypdf>=5.4.0",
    "pypdf2>=3.0.1",
    "requests>=2.32.3",
    "sentence-transformers>=4.1.0",
    "python-dotenv>=1.1.0",
    "openai>=1.75.0",
    "langchain-openai>=0.3.14",
    "langchain-chroma>=0.1.1",
]
